+++ samples/Web/Startup.fs
@@ -21,6 +21,7 @@ type Arguments =
     | [<CliPrefix(CliPrefix.None); Last; AltCommandLine "ms">]  MsSql    of ParseResults<Storage.Sql.Ms.Parameters>
     | [<CliPrefix(CliPrefix.None); Last; AltCommandLine "my">]  MySql    of ParseResults<Storage.Sql.My.Parameters>
     | [<CliPrefix(CliPrefix.None); Last; AltCommandLine "pg">]  Postgres of ParseResults<Storage.Sql.Pg.Parameters>
+    | [<CliPrefix(CliPrefix.None); Last>]                       Mdb      of ParseResults<Storage.MessageDb.Parameters>
     interface IArgParserTemplate with
         member a.Usage = a |> function
             | Verbose ->                    "Include low level Domain and Store logging in screen output."
@@ -34,6 +35,7 @@ type Arguments =
             | MsSql _ ->                    "specify storage in Sql Server (--help for options)."
             | MySql _ ->                    "specify storage in MySql (--help for options)."
             | Postgres _ ->                 "specify storage in Postgres (--help for options)."
+            | Mdb _ ->                      "specify storage in MessageDB (--help for options)."
 module Arguments =
     let parse argv =
         let programName = System.Reflection.Assembly.GetEntryAssembly().GetName().Name
@@ -94,16 +96,18 @@ type Startup() =
             | Memory _ ->
                 log.Fatal("Web App is using Volatile Store; Storage options: {options:l}", options)
                 Storage.MemoryStore.config (), log
+            | Mdb sp ->
+                log.Information("MessageDB Storage options: {options:l}", options)
+                Storage.MessageDb.config log cache sp, log
             | x -> Storage.missingArg (sprintf "unexpected subcommand %A" x)
         Services.register(services, storeConfig, storeLog)
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
     static member Configure(app: IApplicationBuilder, env: IHostEnvironment) : unit =
         if env.IsDevelopment() then app.UseDeveloperExceptionPage() |> ignore
-        else app.UseHsts() |> ignore
+        else app.UseHsts().UseHttpsRedirection() |> ignore
 
         app
-            .UseHttpsRedirection()
             .UseCors(fun x -> x.WithOrigins("https://www.todobackend.com").AllowAnyHeader().AllowAnyMethod() |> ignore)
             .UseRouting()
             .UseEndpoints(fun endpoints -> endpoints.MapControllers() |> ignore) |> ignore