+++ tests/Equinox.EventStore.Integration/Infrastructure.fs
@@ -38,21 +38,12 @@ module SerilogHelpers =
         match evt with
         | Log.WriteSuccess _ -> EsAct.Append
         | Log.WriteConflict _ -> EsAct.AppendConflict
-#if !STORE_EVENTSTOREDB && !STORE_MESSAGEDB // For gRPC, no slice information is available
+#if !STORE_EVENTSTOREDB // For gRPC, no slice information is available
         | Log.Slice (Direction.Forward,_) -> EsAct.SliceForward
         | Log.Slice (Direction.Backward,_) -> EsAct.SliceBackward
 #endif
-#if STORE_MESSAGEDB // MessageDB has no backwards reading
-        | Log.Slice _ -> EsAct.SliceForward
-        | Log.Batch _ -> EsAct.BatchForward
-        // we pretend that reading the last event
-        // is a batched backwards read to comply with
-        // the test harnesses expectations
-        | Log.Last _ -> EsAct.BatchBackward
-#else
         | Log.Batch (Direction.Forward,_,_) -> EsAct.BatchForward
         | Log.Batch (Direction.Backward,_,_) -> EsAct.BatchBackward
-#endif
     let (|EsEvent|_|) (logEvent : LogEvent) : Log.Metric option =
         logEvent.Properties.Values |> Seq.tryPick (function
             | SerilogScalar (:? Log.Metric as e) -> Some e