+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 - `DynamoStore`/`DynamoStore.Prometheus`: Implements the majority of the `CosmosStore` functionality via `FSharp.AWS.DynamoDB` [#321](https://github.com/jet/equinox/pull/321)
 - `EventStoreDb`: As per `EventStore` module, but using the modern `EventStore.Client.Grpc.Streams` client [#196](https://github.com/jet/equinox/pull/196)
 - `eqx dump`: `-s` flag is now optional
+- `MessageDb`: Added a [message-db](https://github.com/message-db/message-db) storage backend
 
 ### Changed